Add monitor.py and Update necessary scripts
